+++ src/main/java/com/plzgraduate/myongjigraduatebe/auth/api/token/TokenApiPresentation.java
@@ -5,7 +5,7 @@ import javax.validation.Valid;
 import org.springframework.web.bind.annotation.RequestBody;
 
 import com.plzgraduate.myongjigraduatebe.auth.api.token.dto.request.TokenRequest;
-import com.plzgraduate.myongjigraduatebe.auth.application.port.in.AccessTokenResponse;
+import com.plzgraduate.myongjigraduatebe.auth.api.token.dto.response.AccessTokenResponse;
 
 import io.swagger.v3.oas.annotations.tags.Tag;
