+++ src/main/java/com/plzgraduate/myongjigraduatebe/auth/api/token/TokenController.java
@@ -7,8 +7,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import com.plzgraduate.myongjigraduatebe.auth.api.token.dto.request.TokenRequest;
-import com.plzgraduate.myongjigraduatebe.auth.application.port.in.AccessTokenResponse;
-import com.plzgraduate.myongjigraduatebe.auth.application.port.in.token.TokenUseCase;
+import com.plzgraduate.myongjigraduatebe.auth.api.token.dto.response.AccessTokenResponse;
+import com.plzgraduate.myongjigraduatebe.auth.application.usecase.token.TokenUseCase;
 import com.plzgraduate.myongjigraduatebe.core.meta.WebAdapter;
 
 import lombok.RequiredArgsConstructor;
