+++ src/main/java/com/plzgraduate/myongjigraduatebe/auth/application/port/FindRefreshTokenPort.java
@@ -1,4 +1,4 @@
-package com.plzgraduate.myongjigraduatebe.auth.application.port.out;
+package com.plzgraduate.myongjigraduatebe.auth.application.port;
 
 import java.util.Optional;
