--- src/main/java/com/plzgraduate/myongjigraduatebe/auth/application/port/in/signin/SignInUseCase.java
@@ -1,7 +1,0 @@
-package com.plzgraduate.myongjigraduatebe.auth.application.port.in.signin;
-
-import com.plzgraduate.myongjigraduatebe.auth.application.port.in.TokenResponse;
-
-public interface SignInUseCase {
-	TokenResponse signIn(SignInCommand signInCommand);
-}
