--- src/main/java/com/plzgraduate/myongjigraduatebe/auth/application/port/in/token/TokenUseCase.java
@@ -1,7 +1,0 @@
-package com.plzgraduate.myongjigraduatebe.auth.application.port.in.token;
-
-import com.plzgraduate.myongjigraduatebe.auth.application.port.in.AccessTokenResponse;
-
-public interface TokenUseCase {
-	AccessTokenResponse createNewToken(TokenCommand tokenCommand);
-}
