+++ src/main/java/com/plzgraduate/myongjigraduatebe/auth/application/service/signin/SignInService.java
@@ -5,10 +5,10 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.plzgraduate.myongjigraduatebe.auth.application.port.in.signin.SignInUseCase;
-import com.plzgraduate.myongjigraduatebe.auth.application.port.in.signin.SignInCommand;
-import com.plzgraduate.myongjigraduatebe.auth.application.port.in.TokenResponse;
-import com.plzgraduate.myongjigraduatebe.auth.application.port.out.SaveRefreshTokenPort;
+import com.plzgraduate.myongjigraduatebe.auth.application.usecase.signin.SignInUseCase;
+import com.plzgraduate.myongjigraduatebe.auth.application.usecase.signin.SignInCommand;
+import com.plzgraduate.myongjigraduatebe.auth.api.signin.dto.response.TokenResponse;
+import com.plzgraduate.myongjigraduatebe.auth.application.port.SaveRefreshTokenPort;
 import com.plzgraduate.myongjigraduatebe.auth.security.JwtAuthenticationToken;
 import com.plzgraduate.myongjigraduatebe.auth.security.TokenProvider;
 import com.plzgraduate.myongjigraduatebe.core.meta.UseCase;
