+++ src/main/java/com/plzgraduate/myongjigraduatebe/auth/application/usecase/signin/SignInUseCase.java
@@ -1,0 +1,7 @@
+package com.plzgraduate.myongjigraduatebe.auth.application.usecase.signin;
+
+import com.plzgraduate.myongjigraduatebe.auth.api.signin.dto.response.TokenResponse;
+
+public interface SignInUseCase {
+	TokenResponse signIn(SignInCommand signInCommand);
+}
