+++ src/main/java/com/plzgraduate/myongjigraduatebe/auth/application/usecase/token/TokenCommand.java
@@ -1,4 +1,4 @@
-package com.plzgraduate.myongjigraduatebe.auth.application.port.in.token;
+package com.plzgraduate.myongjigraduatebe.auth.application.usecase.token;
 
 import lombok.Builder;
 import lombok.Getter;
