+++ src/main/java/com/plzgraduate/myongjigraduatebe/auth/application/usecase/token/TokenUseCase.java
@@ -1,0 +1,7 @@
+package com.plzgraduate.myongjigraduatebe.auth.application.usecase.token;
+
+import com.plzgraduate.myongjigraduatebe.auth.api.token.dto.response.AccessTokenResponse;
+
+public interface TokenUseCase {
+	AccessTokenResponse createNewToken(TokenCommand tokenCommand);
+}
