+++ src/test/java/com/plzgraduate/myongjigraduatebe/auth/application/service/token/TokenServiceTest.java
@@ -13,9 +13,9 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import com.plzgraduate.myongjigraduatebe.auth.application.port.in.AccessTokenResponse;
-import com.plzgraduate.myongjigraduatebe.auth.application.port.in.token.TokenCommand;
-import com.plzgraduate.myongjigraduatebe.auth.application.port.out.FindRefreshTokenPort;
+import com.plzgraduate.myongjigraduatebe.auth.api.token.dto.response.AccessTokenResponse;
+import com.plzgraduate.myongjigraduatebe.auth.application.usecase.token.TokenCommand;
+import com.plzgraduate.myongjigraduatebe.auth.application.port.FindRefreshTokenPort;
 import com.plzgraduate.myongjigraduatebe.auth.security.TokenProvider;
 
 @ExtendWith(MockitoExtension.class)
