+++ src/test/java/com/plzgraduate/myongjigraduatebe/support/WebAdaptorTestSupport.java
@@ -15,8 +15,8 @@ import org.springframework.web.context.WebApplicationContext;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.plzgraduate.myongjigraduatebe.auth.api.signin.SignInController;
 import com.plzgraduate.myongjigraduatebe.auth.api.token.TokenController;
-import com.plzgraduate.myongjigraduatebe.auth.application.port.in.signin.SignInUseCase;
-import com.plzgraduate.myongjigraduatebe.auth.application.port.in.token.TokenUseCase;
+import com.plzgraduate.myongjigraduatebe.auth.application.usecase.signin.SignInUseCase;
+import com.plzgraduate.myongjigraduatebe.auth.application.usecase.token.TokenUseCase;
 import com.plzgraduate.myongjigraduatebe.auth.security.TokenProvider;
 import com.plzgraduate.myongjigraduatebe.core.config.JpaAuditingConfig;
 import com.plzgraduate.myongjigraduatebe.core.config.QuerydslConfig;
